drivers/gpio: put all gpio functionality into imx6ull-gpio.c
first step to add separate implementation files for imx6ull and imxrt106x gpio

JIRA: RTOS-507
